make file hasher easier to re-use
